LayoutTests/imported/w3c/web-platform-tests/css/css-tables/height-distribution/extra-height-given-to-all-row-groups-004-expected.txt


FAIL #theTable 1 assert_equals:
<table id="theTable" style="height:100px">
  <thead data-expected-height="50">
    <tr>
      <td><div></div></td>
    </tr>
  </thead>
  <tbody data-expected-height="50">
    <tr>
      <td><div></div></td>
    </tr>
  </tbody>
</table>
height expected 50 but got 10
